# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "get": {
        "operationId": "PostsController_findAll",
        "summary": "Get all posts",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostsController_create",
        "summary": "Create post",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "summary": "Get post by id",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "summary": "Update post",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "summary": "Delete post",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Post 4 deleted"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/organisers": {
      "get": {
        "operationId": "OrganisersController_findAll",
        "summary": "List all organiser applications",
        "description": "Admin (approval queue) / Owner (read-only stats drill-down).",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "verified",
                "rejected"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/apply": {
      "post": {
        "operationId": "OrganisersController_apply",
        "summary": "Apply to become a Verified Organizer",
        "description": "Creates a pending application. Does not itself grant the organizer role — see PATCH /:id.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyOrganiserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/{id}": {
      "patch": {
        "operationId": "OrganisersController_setStatus",
        "summary": "Verify or reject an organiser application",
        "description": "Admin-only. On verification, also flips the applicant's account role to organizer.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganiserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/profile": {
      "get": {
        "operationId": "OrganisersController_profile",
        "summary": "Get a user's organiser application/plan status",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application status, plan, and plan limits"
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/analytics": {
      "get": {
        "operationId": "OrganisersController_analytics",
        "summary": "Get an organiser's tournament analytics",
        "description": "Aggregates their own events/registrations — same scoping as the Organizer Dashboard.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
    "/api/organisers/subscription/upgrade": {
      "post": {
        "operationId": "OrganisersController_upgradePlan",
        "summary": "Upgrade the organiser subscription tier",
        "description": "Free Organizer → Premium Organizer. Logs a simulated payment.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. apply/profile/analytics/subscription-upgrade: user | organizer | admin. PATCH :id (verify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeOrganiserPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "organisers"
        ]
      }
    },
...
    "/api/ratings": {
      "post": {
        "operationId": "RatingsController_create",
        "summary": "Rate an organiser or moderator",
        "description": "Doc §13 — player ratings feed the organiser/moderator trust score.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Any authenticated role may rate an organiser/moderator they interacted with.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRatingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ratings"
        ]
      },
      "get": {
        "operationId": "RatingsController_forTarget",
        "summary": "Get ratings + average for an organiser or moderator",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. Any authenticated role may rate an organiser/moderator they interacted with.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "organiser",
                "moderator"
              ],
              "type": "string"
            }
          },
          {
            "name": "targetUserId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 6,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Average score, count, and the individual rating records"
          }
        },
        "tags": [
          "ratings"
        ]
      }
    }
  },
  "info": {
    "title": "Gameunity Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | community_manager | moderator | user"
      }
    },
    "schemas": {
...
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Looking for squad"
          },
          "content": {
            "type": "string",
            "example": "Need two players for ranked grind"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "authorId": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "communityId",
          "authorId"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Tournament practice tonight"
          },
          "content": {
            "type": "string",
            "minLength": 8,
            "maxLength": 500,
            "example": "Join us at 9PM for ranked practice sessions"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "authorId": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "title",
          "content",
          "communityId",
          "authorId"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Tournament practice tonight"
          },
          "content": {
            "type": "string",
            "minLength": 8,
            "maxLength": 500,
            "example": "Join us at 9PM for ranked practice sessions"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "authorId": {
            "type": "number",
            "example": 3
          }
        }
      },
...
      "ApplyOrganiserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 4
          },
          "experienceNote": {
            "type": "string",
            "example": "Ran 3 campus BGMI tournaments in the last year."
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateOrganiserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "verified",
            "enum": [
              "verified",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpgradeOrganiserPlanDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 6
          },
          "plan": {
            "type": "string",
            "example": "premium",
            "enum": [
              "free",
              "premium"
            ]
          }
        },
        "required": [
          "userId",
          "plan"
        ]
      },
...
      "CreateRatingDto": {
        "type": "object",
        "properties": {
          "targetType": {
            "type": "string",
            "example": "organiser",
            "enum": [
              "organiser",
              "moderator"
            ]
          },
          "targetUserId": {
            "type": "number",
            "example": 6
          },
          "raterId": {
            "type": "number",
            "example": 4
          },
          "score": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "example": "Ran a smooth, well-organized tournament."
          }
        },
        "required": [
          "targetType",
          "targetUserId",
          "raterId",
          "score"
        ]
      }
    }
  }
}